:truck: api: move get-certificate route and controller from lib to src
