+++ api/tests/certification/results/unit/application/certification-route_test.js
@@ -8,10 +8,21 @@ describe('Certification | Results | Unit | Application | Certification Route', f
 
   beforeEach(async function () {
     sinon.stub(certificationController, 'findUserCertifications').returns('ok');
+    sinon.stub(certificationController, 'getCertification').callsFake((request, h) => h.response('ok').code(200));
     sinon.stub(securityPreHandlers, 'hasAtLeastOneAccessOf').returns(() => true);
 
     httpTestServer = new HttpTestServer();
     await httpTestServer.register(moduleUnderTest);
+  });
+
+  describe('GET /api/certifications/:id', function () {
+    it('should exist', async function () {
+      // when
+      const response = await httpTestServer.request('GET', '/api/certifications/1');
+
+      // then
+      expect(response.statusCode).to.equal(200);
+    });
   });
 
   describe('GET /api/certifications', function () {
