--- api/tests/integration/application/certifications/index_test.js
@@ -1,26 +1,0 @@
-import { certificationController } from '../../../../lib/application/certifications/certification-controller.js';
-import * as moduleUnderTest from '../../../../lib/application/certifications/index.js';
-import { securityPreHandlers } from '../../../../src/shared/application/security-pre-handlers.js';
-import { expect, HttpTestServer, sinon } from '../../../test-helper.js';
-
-describe('Integration | Application | Route | Certifications', function () {
-  let httpTestServer;
-
-  beforeEach(async function () {
-    sinon.stub(certificationController, 'getCertification').callsFake((request, h) => h.response('ok').code(200));
-    sinon.stub(securityPreHandlers, 'hasAtLeastOneAccessOf').returns(() => true);
-
-    httpTestServer = new HttpTestServer();
-    await httpTestServer.register(moduleUnderTest);
-  });
-
-  describe('GET /api/certifications/:id', function () {
-    it('should exist', async function () {
-      // when
-      const response = await httpTestServer.request('GET', '/api/certifications/1');
-
-      // then
-      expect(response.statusCode).to.equal(200);
-    });
-  });
-});
