+++ api/tests/unit/application/certifications/certification-controller_test.js
@@ -3,77 +3,8 @@ import { ChallengeDeneutralized } from '../../../../lib/domain/events/ChallengeD
 import { ChallengeNeutralized } from '../../../../lib/domain/events/ChallengeNeutralized.js';
 import { usecases } from '../../../../lib/domain/usecases/index.js';
 import { domainBuilder, expect, hFake, sinon } from '../../../test-helper.js';
-import { getI18n } from '../../../tooling/i18n/i18n.js';
 
 describe('Unit | Controller | certifications-controller', function () {
-  describe('#getCertification', function () {
-    it('should return a serialized private certificate given by id', async function () {
-      // given
-      const userId = 1;
-      const certificationId = 2;
-      const request = {
-        auth: { credentials: { userId } },
-        params: { id: certificationId },
-        i18n: getI18n(),
-      };
-      const locale = 'fr-fr';
-      const requestResponseUtilsStub = { extractLocaleFromRequest: sinon.stub() };
-      const privateCertificate = domainBuilder.buildPrivateCertificate.validated({
-        id: certificationId,
-        firstName: 'Dorothé',
-        lastName: '2Pac',
-        birthdate: '2000-01-01',
-        birthplace: 'Sin City',
-        isPublished: true,
-        date: new Date('2020-01-01T00:00:00Z'),
-        deliveredAt: new Date('2021-01-01T00:00:00Z'),
-        certificationCenter: 'Centre des choux de Bruxelles',
-        pixScore: 456,
-        commentForCandidate: 'Cette personne est impolie !',
-        certifiedBadgeImages: [],
-        verificationCode: 'P-SUPERCODE',
-        maxReachableLevelOnCertificationDate: 6,
-      });
-      sinon.stub(usecases, 'getPrivateCertificate');
-      usecases.getPrivateCertificate.withArgs({ userId, certificationId, locale }).resolves(privateCertificate);
-      requestResponseUtilsStub.extractLocaleFromRequest.withArgs(request).returns(locale);
-
-      // when
-      const response = await certificationController.getCertification(request, hFake, {
-        requestResponseUtils: requestResponseUtilsStub,
-      });
-
-      // then
-      expect(response).to.deep.equal({
-        data: {
-          id: '2',
-          type: 'certifications',
-          attributes: {
-            'first-name': 'Dorothé',
-            'last-name': '2Pac',
-            birthdate: '2000-01-01',
-            birthplace: 'Sin City',
-            'certification-center': 'Centre des choux de Bruxelles',
-            date: new Date('2020-01-01T00:00:00Z'),
-            'delivered-at': new Date('2021-01-01T00:00:00Z'),
-            'is-published': true,
-            'pix-score': 456,
-            status: 'validated',
-            'comment-for-candidate': 'Cette personne est impolie !',
-            'certified-badge-images': [],
-            'verification-code': 'P-SUPERCODE',
-            'max-reachable-level-on-certification-date': 6,
-          },
-          relationships: {
-            'result-competence-tree': {
-              data: null,
-            },
-          },
-        },
-      });
-    });
-  });
-
   describe('#getCertificationByVerificationCode', function () {
     it('should return a serialized shareable certificate given by verification code', async function () {
       // given
